// lib/seo/combo-data.generated.ts
export type ComboTopicKey = 'mingGong' | 'fuQi' | 'guanLu' | 'caiBo';

export interface ComboTopicContent {
  dingdiao: string;
  lundian: string;
  yiju: string;
  chuchu: string;
}